Some refactoring, and fixing source download
